Move region and store from common to store crate
