+++ src/pd/retry.rs
@@ -2,6 +2,8 @@
 
 //! A utility module for managing and retrying PD requests.
 
+use crate::{Region, RegionId, StoreId};
+
 use async_trait::async_trait;
 use futures_timer::Delay;
 use grpcio::Environment;
@@ -11,9 +13,7 @@ use std::{
     sync::Arc,
     time::{Duration, Instant},
 };
-use tikv_client_common::{
-    security::SecurityManager, stats::pd_stats, Error, Region, RegionId, Result, StoreId, Timestamp,
-};
+use tikv_client_common::{security::SecurityManager, stats::pd_stats, Error, Result, Timestamp};
 use tikv_client_pd::cluster::{Cluster, Connection};
 use tokio::sync::RwLock;
