+++ tikv-client-store/src/lib.rs
@@ -4,16 +4,21 @@
 extern crate log;
 
 mod errors;
+pub mod region;
+mod store_builder;
 
-pub use self::errors::{HasError, HasRegionError};
+pub use crate::errors::{HasError, HasRegionError};
+#[doc(inline)]
+pub use crate::region::{Region, RegionId, RegionVerId, StoreId};
+#[doc(inline)]
+pub use crate::store_builder::StoreBuilder;
+pub use kvproto::tikvpb::TikvClient;
+
 use derive_new::new;
 use futures::{future::BoxFuture, prelude::*};
 use grpcio::{CallOption, Environment};
-pub use kvproto::tikvpb::TikvClient;
 use std::{sync::Arc, time::Duration};
-use tikv_client_common::{
-    security::SecurityManager, stats::tikv_stats, ErrorKind, Region, Result, StoreBuilder,
-};
+use tikv_client_common::{security::SecurityManager, stats::tikv_stats, ErrorKind, Result};
 
 /// A trait for connecting to TiKV stores.
 pub trait KvConnect: Sized + Send + Sync + 'static {
